# proto/rpc.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/rpc.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/bucket/create": {
      "post": {
        "operationId": "CreateBucket",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protoCreateBucketResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoCreateBucketRequest"
            }
          }
        ],
        "tags": [
          "StoreApiService"
        ]
      }
    },
    "/config/download": {
      "get": {
        "operationId": "GetConfigStoreInfo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protoGetConfigStoreInfoResponse"
            }
          }
        },
        "tags": [
          "StoreApiService"
        ]
      }
    },
    "/config/upload": {
      "post": {
        "operationId": "UpdateConfigStoreInfo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protoUpdateConfigStoreInfoResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoUpdateConfigStoreInfoRequest"
            }
          }
        ],
        "tags": [
          "StoreApiService"
        ]
      }
    },
    "/object/download": {
      "post": {
        "operationId": "GetObject",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protoGetObjectResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoGetObjectRequest"
            }
          }
        ],
        "tags": [
          "StoreApiService"
        ]
      }
    },
    "/object/downloadurl": {
      "post": {
        "operationId": "GetObjectPresignedUrl",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protoGetObjectPresignedUrlResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoGetObjectPresignedUrlRequest"
            }
          }
        ],
        "tags": [
          "StoreApiService"
        ]
      }
    },
    "/object/upload": {
      "post": {
        "operationId": "PutObject",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protoPutObjectResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoPutObjectRequest"
            }
          }
        ],
        "tags": [
          "StoreApiService"
        ]
      }
    },
    "/object/uploadurl": {
      "post": {
        "operationId": "PutObjectPresignedUrl",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protoPutObjectPresignedUrlResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoPutObjectPresignedUrlRequest"
            }
          }
        ],
        "tags": [
          "StoreApiService"
        ]
      }
    }
  },
  "definitions": {
    "protoConfigInfo": {
      "type": "object",
      "properties": {
        "accesskey": {
          "type": "string"
        },
        "secretkey": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "endpoint": {
          "type": "string"
        }
      }
    },
    "protoCreateBucketRequest": {
      "type": "object",
      "properties": {
        "bucketname": {
          "type": "string"
        },
        "publicread": {
          "type": "string"
        }
      }
    },
    "protoCreateBucketResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      }
    },
    "protoGetConfigStoreInfoResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/protoConfigInfo"
        }
      }
    },
    "protoGetObjectPresignedUrlRequest": {
      "type": "object",
      "properties": {
        "bucketname": {
          "type": "string"
        },
        "key": {
          "type": "string"
        },
        "expiretime": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protoGetObjectPresignedUrlResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      }
    },
    "protoGetObjectRequest": {
      "type": "object",
      "properties": {
        "bucketname": {
          "type": "string"
        },
        "key": {
          "type": "string"
        },
        "contenttype": {
          "type": "string"
        }
      }
    },
    "protoGetObjectResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      }
    },
    "protoPutObjectPresignedUrlRequest": {
      "type": "object",
      "properties": {
        "bucketname": {
          "type": "string"
        },
        "key": {
          "type": "string"
        },
        "publicread": {
          "type": "string"
        },
        "contenttype": {
          "type": "string"
        },
        "contentmaxlength": {
          "type": "string",
          "format": "int64"
        },
        "expiretime": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protoPutObjectPresignedUrlResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      }
    },
    "protoPutObjectRequest": {
      "type": "object",
      "properties": {
        "bucketname": {
          "type": "string"
        },
        "key": {
          "type": "string"
        },
        "publicread": {
          "type": "string"
        },
        "contenttype": {
          "type": "string"
        },
        "contentmaxlength": {
          "type": "string",
          "format": "int64"
        },
        "filecontent": {
          "type": "string"
        },
        "expiretime": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protoPutObjectResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      }
    },
    "protoUpdateConfigStoreInfoRequest": {
      "type": "object",
      "properties": {
        "domain": {
          "type": "string"
        },
        "accesskey": {
          "type": "string"
        },
        "secretkey": {
          "type": "string"
        },
        "region": {
          "type": "string"
        }
      }
    },
    "protoUpdateConfigStoreInfoResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      }
    }
  }
}